Коммит содержит изменения работы кнопок (не всех, но большенство) под zenject. Беру во внимание тот фатк что скрипт InputHandles работает некоректно по отношению к нескольким сценам, пользующимся им. Как следствие было создано 2 копии данного скритпа StartSceneInputHandels MenuInputHandels, хранящие в себе толко те элементы UI, которые используются только в конкретной сцене (соответсвенно название скриптов соответсвующее). Так же была добавлена функция востонавления пароля пользователя и существенно пределан скрипт AuthorizationProcessing под "более удобную для понимания" семантику кода скрипта (Однако этот скрипт следует удалить и подумать над более правильной сетевой системой в дальнейшем). :)
