+++ Assets/Editor/SpineSettings.asset.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 680b999a8c7e74b489289916f6a58224
-PrefabImporter:
+guid: 4874bf2a4b733f6488bbe872e41e37ec
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
