+++ Assets/Plugins/Zenject/Source/Runtime/Kernels/MonoKernel.cs
@@ -7,14 +7,15 @@ namespace Zenject
 {
     public abstract class MonoKernel : MonoBehaviour
     {
+#pragma warning disable 0649
         [InjectLocal]
-        TickableManager _tickableManager = null;
+        TickableManager _tickableManager;
 
         [InjectLocal]
-        InitializableManager _initializableManager = null;
+        InitializableManager _initializableManager;
 
         [InjectLocal]
-        DisposableManager _disposablesManager = null;
+        DisposableManager _disposablesManager;
 
         bool _hasInitialized;
         bool _isDestroyed;
