+++ Assets/Scripts/Installers/MenuInstaller.cs
@@ -7,11 +7,11 @@ namespace LWT.Installers
     public class MenuInstaller : MonoInstaller
     {
         [SerializeField]
-        private InputHandles.UIElements uIElements;
+        private MenuInputHandels.MenuUIElements uIElements = null;
 
         public override void InstallBindings()
         {
-            Container.BindInterfacesAndSelfTo<InputHandles>().AsSingle();
+            Container.BindInterfacesAndSelfTo<MenuInputHandels>().AsSingle();
             Container.BindInstance(uIElements);
         }
     }
