+++ Assets/Scripts/Installers/StartInstaller.cs
@@ -1,0 +1,19 @@
+﻿using LWT.System;
+using UnityEngine;
+using Zenject;
+
+namespace LWT.Installers
+{
+    public class StartInstaller : MonoInstaller
+    {
+        [SerializeField]
+        private StartSceneInputHandels.StartUIElements uIElements = null;
+
+        public override void InstallBindings()
+        {
+            Container.BindInterfacesAndSelfTo<StartSceneInputHandels>().AsSingle();
+            Container.BindInstance(uIElements);
+        }
+    }
+}
+
