+++ Assets/Scripts/Networking/Authorization.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+
+namespace LWT.Networking.Authorization
+{
+    public class Authorization : MonoBehaviour
+    {
+
+    }
+}
+
