+++ Assets/Scripts/Networking/Authorization.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 702ab1b9fab6d5246bc3277c4d8293c0
+guid: 3f5ba69a3f38a294f8bc49cefacc3356
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
