+++ Assets/Scripts/SetEmailToGlobalBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a2049db4638f3d40a79576b102f2806
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
