+++ Assets/Scripts/System/MenuInputHandels.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: abe8c2e463a12e646a065429f5f072ec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
