+++ Assets/Scripts/System/StartSceneInputHandels.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3bae26db8320fe140b09094f2a909b8f
+guid: 0f1e92d51e2799d4bb8c6bacc72f4404
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
