+++ Assets/Scripts/UI/StartScene/AuthorizationPanel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cbd777ab4a49b194baa7868c1a8fd40f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
