+++ Assets/Scripts/UI/StartScene/ContextPanel.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+
+namespace LWT.UI.StartScene
+{
+    public class ContextPanel : MonoBehaviour
+    {
+
+        private void Start()
+        {
+
+        }
+    }
+}
