+++ Assets/Scripts/UI/StartScene/ContextPanel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9a434592c579f8a47862f501477587a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
