+++ Assets/Scripts/UI/StartScene/RegistrationPanel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b8ae6797b7683914cbe07261f226d6ab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
