+++ Assets/Scripts/UI/StartScene/RestorePasswordPanel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 17b15eedf5811824194ad15e89ee4782
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
